+++ src/LibHac/Fs/DirectoryEntry.cs
@@ -1,26 +1,8 @@
 ﻿using System;
 using System.Runtime.InteropServices;
 using LibHac.Common;
-using LibHac.FsSystem;
 
 namespace LibHac.Fs;
-
-public class DirectoryEntryEx
-{
-    public string Name { get; set; }
-    public string FullPath { get; set; }
-    public NxFileAttributes Attributes { get; set; }
-    public DirectoryEntryType Type { get; set; }
-    public long Size { get; set; }
-
-    public DirectoryEntryEx(string name, string fullPath, DirectoryEntryType type, long size)
-    {
-        Name = name;
-        FullPath = PathTools.Normalize(fullPath);
-        Type = type;
-        Size = size;
-    }
-}
 
 [StructLayout(LayoutKind.Explicit)]
 public struct DirectoryEntry
