+++ src/LibHac/Tools/FsSystem/AesXtsFileSystem.cs
@@ -4,6 +4,7 @@ using LibHac.Common;
 using LibHac.Fs;
 using LibHac.Fs.Fsa;
 using LibHac.FsSystem;
+using LibHac.Tools.Fs;
 using LibHac.Util;
 
 namespace LibHac.Tools.FsSystem;
