+++ src/LibHac/Tools/FsSystem/FileSystemExtensions.cs
@@ -7,6 +7,7 @@ using LibHac.Common;
 using LibHac.Fs;
 using LibHac.Fs.Fsa;
 using LibHac.FsSystem;
+using LibHac.Tools.Fs;
 using LibHac.Util;
 using Path = LibHac.Fs.Path;
