+++ src/LibHac/Tools/FsSystem/RomFs/RomFsBuilder.cs
@@ -6,6 +6,7 @@ using LibHac.Common;
 using LibHac.Fs;
 using LibHac.Fs.Fsa;
 using LibHac.FsSystem;
+using LibHac.Tools.Fs;
 using LibHac.Util;
 
 namespace LibHac.Tools.FsSystem.RomFs;
