+++ src/LibHac/Tools/FsSystem/Save/SaveDataFileSystemCore.cs
@@ -2,6 +2,7 @@
 using LibHac.Common;
 using LibHac.Fs;
 using LibHac.Fs.Fsa;
+using LibHac.Tools.Fs;
 using LibHac.Util;
 using Path = LibHac.Fs.Path;
