+++ src/hactoolnet/ProcessNca.cs
@@ -7,6 +7,7 @@ using LibHac.Fs.Fsa;
 using LibHac.Fs.Impl;
 using LibHac.FsSystem;
 using LibHac.Npdm;
+using LibHac.Tools.Fs;
 using LibHac.Tools.FsSystem;
 using LibHac.Tools.FsSystem.NcaUtils;
 using static hactoolnet.Print;
