+++ src/hactoolnet/ProcessSave.cs
@@ -9,6 +9,7 @@ using LibHac.Fs;
 using LibHac.Fs.Fsa;
 using LibHac.Fs.Impl;
 using LibHac.FsSystem;
+using LibHac.Tools.Fs;
 using LibHac.Tools.FsSystem;
 using LibHac.Tools.FsSystem.Save;
 using static hactoolnet.Print;
